Refactor based on learning
The original project was an attempt to see what
can be done with guice and guava. The main idea
was that an API would be exposed which guice would
bind into. Then the API would be called per package
and post events on state changes via the event bus.
Then any package listening to the event bus would
be able to respond. Finally this was tied in
together with cucumber.

However as things progressed the learning evolved
and refactoring was attempted. It was just painful
to see the over-engineered solution. So this is a
complete cleanup. The basics have been added.

Functionality is still to come.

Origin Branch: refactor/clean